Avoid duplicating crowdanki_uuid when cloning deck config
Fix #118.

Testing, it seems that I somehow haven't broken the tests despite
fiddling with anki/hook_vendor without adding to the overrides.

Note that deck_conf_did_add_config is not available as a "legacy"
hook, so the new hook mechanism has to be used.  It might (or might
not) be worth converting the existing hooks to the new mechanism.